+++ packages/template-retail-react-app/app/pages/checkout/partials/contact-info.jsx
@@ -46,6 +46,7 @@ const ContactInfo = () => {
     const login = useAuthHelper(AuthHelpers.LoginRegisteredUserB2C)
     const logout = useAuthHelper(AuthHelpers.Logout)
     const updateCustomerForBasket = useShopperBasketsMutation('updateCustomerForBasket')
+    const transferBasket = useShopperBasketsMutation('transferBasket')
 
     const {step, STEPS, goToStep, goToNextStep} = useCheckout()
 
@@ -69,6 +70,12 @@ const ContactInfo = () => {
                 })
             } else {
                 await login.mutateAsync({username: data.email, password: data.password})
+
+                // Because we lazy load the basket there is no guarantee that a basket exists for the newly registered
+                // user, for this reason we must transfer the ownership of the previous basket to the logged in user.
+                await transferBasket.mutateAsync({
+                    parameters: {overrideExisting: true}
+                })
             }
             goToNextStep()
         } catch (error) {